Fix bug on calling to_peer_paserk_id on v2/4 local key.
